refactor: move credential_type to a unified module
